Remove 2 GiB size limit for downloads.
.pkz files larger than 2 GiB are supported now.

+++ src/client/client.h
@@ -440,7 +440,7 @@ typedef struct client_static_s {
         int         pending;            // number of non-finished entries in queue
         dlqueue_t   *current;           // path being downloaded
         int         percent;            // how much downloaded
-        int         position;           // how much downloaded (in bytes)
+        int64_t     position;           // how much downloaded (in bytes)
         qhandle_t   file;               // UDP file transfer from server
         char        temp[MAX_QPATH + 4];// account 4 bytes for .tmp suffix
 #if USE_ZLIB
